feat(posts): multi-image post wiring (up to 5 images)
PostImageCarousel in detail/profile/discover, native multi-select
pick + 4:3 crop each in create/edit, image_urls array + first-image
mirror handling in PostService. Migration 0027 committed previously.

+++ lib/widgets/post_image_carousel.dart
@@ -7,7 +7,7 @@ import '../theme.dart';
 /// the feed card and post detail.
 class PostImageCarousel extends StatefulWidget {
   final List<String> imageUrls;
-  final int decodeWidth;
+  final double decodeWidth;
 
   const PostImageCarousel({
     super.key,
